@@ -1,28 +1,25 @@
 {
-  "name": "minibook-example",
-  "private": true,
-  "version": "0.9.3",
-  "author": "sunflowerdeath",
-  "license": "Unlicense",
-  "devDependencies": {
-    "@babel/core": "^7.0.0",
-    "@babel/preset-env": "^7.0.0",
-    "@babel/preset-react": "^7.0.0",
-    "babel-loader": "^8.0.2",
-    "html-webpack-plugin": "^3.2.0",
-    "minimark-loader": "^0.9.3",
-    "webpack": "^4.12.1",
-    "webpack-cli": "^3.1.0",
-    "webpack-dev-server": "^3.11.0"
-  },
-  "dependencies": {
-    "minibook": "^0.9.3",
-    "minimark-renderer": "^0.9.3",
-    "prop-types": "^15.6.0"
-  },
-  "scripts": {
-    "start": "webpack-dev-server",
-    "build": "NODE_ENV=production webpack",
-    "analyze": "NODE_ENV=production webpack --profile --json > stats.json && webpack-bundle-analyzer ./stats.json"
-  }
+	"name": "minibook-example",
+	"private": true,
+	"version": "0.9.3",
+	"author": "sunflowerdeath",
+	"license": "Unlicense",
+	"devDependencies": {
+		"esbuild-loader": "^4.0.2",
+		"minimark-loader": "*",
+		"raw-loader": "^4.0.2"
+	},
+	"dependencies": {
+		"html-webpack-plugin": "^5.5.3",
+		"minibook": "*",
+		"minimark-renderer": "*",
+		"prop-types": "^15.8.1",
+		"webpack": "^5.89.0",
+		"webpack-dev-server": "^4.15.1"
+	},
+	"scripts": {
+		"start": "webpack-dev-server",
+		"build": "NODE_ENV=production webpack",
+		"analyze": "NODE_ENV=production webpack --profile --json > stats.json && webpack-bundle-analyzer ./stats.json"
+	}
 }
